feat: implemented RecyclerView in the HomeFragment
